Move createAllResiduesFromDsspData and remaining methods to DSSP Parser
