+++ plcc/src/io/DsspParser.java
@@ -9,7 +9,12 @@
 package io;
 
 // imports
+import static io.FileParser.s_sulfurBridges;
 import java.util.ArrayList;
+import java.util.HashMap;
+import plcc.Settings;
+import proteinstructure.AminoAcid;
+import proteinstructure.Residue;
 import proteinstructure.SSE;
 import tools.DP;
 
@@ -30,6 +35,8 @@ public class DsspParser {
     static Integer curLineNumDSSP = null;
     static String curLineDSSP = null;
     static Integer curResNumDSSP = null;
+    
+    static boolean dataInitDone = false;
     
     
     /**
@@ -180,6 +187,197 @@ public class DsspParser {
         }
         else {
             return(false);
+        }
+    }
+    
+    
+    /**
+     * Parses the DSSP data and creates the residue list from it.
+     * @param isCIF true if using mmCIF parser and mmCIF pdb file as chain IDs may be 4 character long then
+     */
+    private static void createAllResiduesFromDsspData(Boolean isCIF) {
+
+        String dLine;
+        Integer dLineNum, dsspResNum, pdbResNum, resIndex, acc;
+        dLineNum = dsspResNum = pdbResNum = 0;
+        String dsspChainID, resName1Letter, sseString, iCode;
+        dsspChainID = resName1Letter = iCode = "";
+        sseString = "UNDEF";
+        Character lastChar = null;
+        Float phi = 0.0f;
+        Float psi = 0.0f;
+        int offset;  // if > 99,999 residues everything is shifted
+
+        s_sulfurBridges = new HashMap<Character, ArrayList<Integer>>();
+        s_interchainSulfurBridges = new HashMap<Character, ArrayList<Integer>>();
+        s_interchainSulfurBridgesChainID = new  HashMap<Character, String>();
+        Residue r;
+
+        for(Integer i = dsspDataStartLine - 1; i < dsspLines.size(); i++) {
+            dLine = dsspLines.get(i);
+            dLineNum = i + 1;
+            
+            offset = Math.max(dLine.split(" ")[0].length() - 5, 0);  // typically first 5 columns code for res num, but if exceeded add offset
+
+            if(dLine.substring(13 + offset, 14 + offset).equals("!")) {       // chain brake
+                if(! FileParser.silent) {
+                    if (! Settings.getBoolean("plcc_B_no_chain_break_info")) {
+                        System.out.println("    DSSP: Found chain brake at DSSP line " + dLineNum + ".");
+                    }
+                }
+            }
+            else {          // parse the residue line
+
+                try {
+                    // column 0 is ignored: blank
+                    dsspResNum = Integer.valueOf(dLine.substring(1 + offset, 5 + offset).trim());
+                    
+                    // last used DSSP res num is later needed for ligands
+                    // (and we only wand to go through dssp file once)
+                    if (isCIF) {
+                        // lets hope dssp res num always increases
+                        lastUsedDsspNum = dsspResNum;
+                    }
+                    
+                    // 5 is ignored: blank
+                    pdbResNum = Integer.valueOf(dLine.substring(6 + offset, 10 + offset).trim());
+                    iCode = dLine.substring(10 + offset, 11 + offset);                    
+                    // with PDB mmCIF files things got more difficult: 4-character chain ids
+                    //     prioritize AUTHCHAIN > CHAIN
+                    if (! isCIF) {
+                        dsspChainID = dLine.substring(11 + offset, 12 + offset);
+                    } else {
+                        dsspChainID = dLine.substring(159 + offset, 163 + offset).trim(); // AUTHCHAIN column 160-163
+                    }
+                    
+                    // 12 is ignored: blank
+                    resName1Letter = dLine.substring(13 + offset, 14 + offset);
+                    // 14+15 are ignored: blank
+                    sseString = dLine.substring(16 + offset, 17 + offset);
+                    // lots of stuff is ignored here
+                    phi = Float.valueOf(dLine.substring(103 + offset, 108 + offset).trim());  // phi backbone angle
+                    psi = Float.valueOf(dLine.substring(109 + offset, 114 + offset).trim());  // psi backbone angle
+                    acc = Integer.valueOf(dLine.substring(35 + offset, 38 + offset).trim());  // solvent accessible surface
+                    // rest is ignored: not needed
+                } catch(Exception e) {
+                    System.err.println("ERROR: Parsing of DSSP line " + dLineNum + " failed: '" + e.getMessage() + "'. DSSP file looks broken.");
+                    //e.printStackTrace();
+                    continue;
+                }
+
+                // successfully parsed, create Residue and add all the info we know atm
+                r = new Residue(pdbResNum, dsspResNum);
+                //System.err.println("DEBUG: Created residue with PDB# " + pdbResNum + ", DSSP# " + dsspResNum + ".");
+                //System.err.println("DEBUG: Residue: " + r.getFancyName() + ".");
+                r.setChainID(dsspChainID);
+                // don't set the chain here, we'll set it later with the pdbChainID while creating the atoms
+                // EDIT: Why not?! I guess I thought that DSSP and PDB files used different chain IDs for residues when I wrote that.
+                // EDIT jnw: do the matching in CIF files later so that each file is only read once
+                if (! isCIF) {
+                    getChainByPdbChainID(dsspChainID).addMolecule(r);
+                }
+
+
+                r.setType(Residue.RESIDUE_TYPE_AA);                   // DSSP files only contain protein residues, ligands are ignored
+                r.setPhi(phi);
+                r.setPsi(psi);
+                r.setAcc(acc);
+
+                // Cysteins which form a sulfur bridge are arbitrary lowercase letters in dssp. A residue
+                // with name 'a' means it is a cysteine, but it has forms a sulfur bridge with a second residue in the file that is
+                // also labeled 'a'.
+                try {
+                    char c = (resName1Letter.toCharArray())[0];
+                    if( ! Character.isUpperCase(c)) {
+                        resName1Letter = "C";   // change residue code cysteine
+
+                        // now go save the sulfur bridge
+                        if(s_sulfurBridges.containsKey(c)) {
+                            // the sulfur bridge partner is already in there
+                            (s_sulfurBridges.get(c)).add(dsspResNum);
+
+                            // Check whether its a interhcain sulfur bridge (different chain IDs)
+                            if(! s_interchainSulfurBridgesChainID.get(c).equals(dsspChainID)) {
+                                ArrayList<Integer> tmpInterchain = new ArrayList<Integer>();
+
+                                // Get the dsspResNum from the first residue of this interchain sulfur bridge
+                                tmpInterchain.add(s_sulfurBridges.get(c).get(0));
+                                // Also add the dsspResNum of the current (second) residue.
+                                tmpInterchain.add(dsspResNum);
+                                s_interchainSulfurBridges.put(c, tmpInterchain);
+                            }
+
+
+                        } else {
+                            // this is the first residue of the sulfur bridge pair
+                            ArrayList<Integer> tmp = new ArrayList<Integer>();
+                            tmp.add(dsspResNum);
+                            s_sulfurBridges.put(c, tmp);
+
+                            // If its the first residue of the sulfur bridge save the chain this residue belongs to.
+                            // This will be used to check if the second residue is on the same change or not.
+                            s_interchainSulfurBridgesChainID.put(c, dsspChainID);
+                        }
+                    }
+                } catch (Exception e) {
+                    DP.getInstance().w("Something is fishy with residue in line " + dLineNum + " of the DSSP file: '" + e.getMessage() + "'. Sulfur bridge trouble?");
+                }
+
+
+                // Fix DSSP file like the one for 2ZW3.pdb which list cysteine residues as 'o' instead of 'c'.
+                if(resName1Letter.equals("O") || resName1Letter.equals("U")) {
+                    DP.getInstance().w("Turned AA with 1-letter code '" + resName1Letter + "' into 'C'.");
+                    resName1Letter = "C";
+                }
+
+                // Define "asparagine or aspartic acid" as asparagine
+                if(resName1Letter.equals("B")) {
+                    DP.getInstance().w("Turned AA with 1-letter code '" + resName1Letter + "' into 'N'.");
+                    resName1Letter = "N";
+                }
+
+                r.setiCode(iCode);
+                r.setAAName1(resName1Letter);
+                r.setName3(AminoAcid.name1ToName3(resName1Letter));
+                r.setSSEString(sseString);
+                // Note: the SSE itself will be set by the getAllSSEs*() functions, it is still NULL atm
+
+                // set model id to '1'. This is always true because DSSP can't handle models and the input files have to be split for it, thus every PDB file it reads has only 1 model
+                r.setModelID("1");
+
+
+                // DEBUG 3eoj
+                //System.err.println("DEBUG --- residue state ---");
+                //if(r.getPdbResNum() == 8 || r.getPdbResNum() == 7 || r.getPdbResNum() == 213) {
+                //    System.err.println("DEBUG: RESIDUE: " + r.getFancyName() + ".");
+                //}
+
+                // add to list of Residues
+                s_molecules.add(r);
+                resIndex = s_molecules.size() - 1;
+                // add index to s_residueIndices
+                s_residueIndices.add(resIndex);
+                // produces null ponter exception in CIF parser and I dont see where we need it (maybe I'll understand later)
+                if (! isCIF) {
+                    resIndexDSSP[dsspResNum] = resIndex;
+                }
+                //resIndexPDB[pdbResNum] = resIndex;  // This will crash because some PDB files use negative residue numbers, omfg.
+                //System.out.println("    DSSP: Added residue PDB # " +  pdbResNum + ", DSSP # " + dsspResNum + " to s_residues at index " + resIndex + ".");
+
+
+            }
+        }
+    }
+    
+    
+        public static ArrayList<String> getDsspLines() {
+        if(dataInitDone) {
+            return(dsspLines);
+        }
+        else {
+            System.err.println("ERROR: FileParser.getDsspLines(): Request for data before initData() was called.");
+            System.exit(1);
+            return(null);
         }
     }
